Separate HF checkpoint test into its own file + move common fixtures to conftest
