+++ python/ray/train/BUILD
@@ -186,6 +186,14 @@ py_test(
 )
 
 py_test(
+    name = "test_huggingface_checkopint",
+    size = "small",
+    srcs = ["tests/test_huggingface_checkpoint.py"],
+    tags = ["team:ml", "exclusive", "ray_air"],
+    deps = [":train_lib"]
+)
+
+py_test(
     name = "test_huggingface_predictor",
     size = "large",
     srcs = ["tests/test_huggingface_predictor.py"],
